fix config.py
had to start allowing any, we can eventually fix this but we would need to define
tons of TF types that are not implemented in the lib we are using.

I also moved the lib into our repo because adding new folders & files
didn't seem to be picked up well and we will be doing that a lot for a while.