Remove PartyDB References
As a part of moving back over to Morten's repository we need to remove
the references to PartyDB.

The only reference remaining is in the README.md which relates to a
stubs repository we've setup. We need to decide what to do with that
repository when we open our request back to Morten's.

Related partydb/planning#1