Dropout Labs Inc., Contributions (#91)
* beginning code of reading and converting protobuf graph to mpc graph

* convert example

* dict cleanup

* mpc graph is being built

* fix merge conflicts

* remove pond changes

* convert unit test

* cleanup

* fix syntax

* slight fixes, renaming conv2d_test

* fix convert example

* forgotten test file

* shape op

* Merge remote-tracking branch 'origin/master' into pooling

* fix Pond context manager

* revert config in convolution_test

* improve typing

* use input providers when running conversion

* pooling wireframe

* strided slice convert

* public pooling

* stride slice convert

* reshape method for public avgpool2d, fix pooling test

* Add node support for Ass

* add support for sub

* implement sum method in Pond*Tensor

* add support for Mul

* abstracted core logic of avgpool2d, implement avgpool2d for private

* Fix tests

* Remove too string typing

* Refactor core.layer

* Add a doc on abstractmethod get_output_shape

* Fix the reshape layer

* wip pooling

* fixed inputs

* wip

* Add Dockerfile to Repository

A simple Dockerfile for containerizing tf-encrypted.

* Update setup.py to fix semver & tf versions

- The semver was invalid (should be `-rc`)
- Updated tensorflow requirements to the latest stable version (1.10.0)

* Introduce bin/serve for running a node

I've introduced bin/serve, a quick and simple script for running a node
compute a result using tf-encrypted.

Related #22

* Introduce docker-compose.yml

We can simply use docker compose to run a remote cluster locally as
standalone processes as if they were deployed in a production setting.

Related #22

* Merge branch 'public-pooling' of github.com:partydb/tf-encrypted into pooling

* finish reshape-style pooling, refactor tests

* Enforce linting on the tests directory

* transpose convert

* single char typo

* reshape convert

* fix lint

* remove padding style check

* Add a first example of using TF for prediction

* Add linting to examples/api folder

* support converting nhwc cond2d

* move channels_last transpose into conv2d layer

* raise exception

* Implement a polling mechanism for retrieving predictions results

* wip

* fix add

* fix sub

* fix mul

* fix convert_test.py

* forgotten self

* convert rsqrt plus batchnorm test

* fix convert test

* Add Makefile to Repository

This commit introduces a Makefile and adopts a Docker continuous
integration pipeline when tf-encrypted is released.

The Makefile also introduces run commands for testing (e.g. `make
test`).

Related #12

* multiplaceholder wip

* working

* add expand dim to pond

* add expand dims to int100

* add expand_dims to converter

* fix lint

* Add setup.cfg to configure pycodestyle

* Add a simple python client

* Clean files

* print shape rather than trying to decode something we cannot

closes #37

* lint layers

* lint protocol dir

* lint crt.py

* added tox.ini for linting

* lint top level files

* support multiple placeholders in an imported model plus stack/pack support

* lint pond.py

* finish linting

* run for all files

* Fix valid average pooling

* remove venv

* ignore ./venv

* make changes to support pycodestyle 2.4

* fix comment

* avgpool convert

* squeeze working

* fix lint

* WIP: update the current work to be docker friendly

* WIP: more work on docker

* decoding for tf tensors

* skin cancer demo

* update

* specify the model with cli

* WIP TODO:Fix session hanging

* move skin cancer demo file to generic model runner helper script

* update help

* Fix TF hanging

* Improve the client API and clean files

* update ReLU coefs for approximation on [-9, 9]

* Document Release Process

I've added a small markdown file which documents our release process for
tf-encrypted.

Closes #74

* Add writer.close to run funtion so we ensure all runs a re dumped to the filesystem

* Roll back to best practice new style

* Refactor the monitor flag

* Standardize naming of "weights_provider"

Internally we were calling the node the "weights_provider" but the serve
script is expecting the value to be provided as the "weight_provider".
I've gone ahead and brought this script in line with the rest of the
system.

* Clean stub files

* Fix docker-compose env variables and image ref

Fix up the forgotten environment variables inside the docker compose
file and refer to the image built by our Makefile!

* Allow api port to be configurable

We need to be able to configure the port that the flask application
listens on! I've gone ahead and updated bin/serve and the docker-compose
config to use the `TFE_API_PORT` environment variable to determine which
port the service should listen on.

* Rename LICENSE and add NOTICE file

We've renamed the license file from `LICENSE-APACHE` to `LICENSE` so
it's in the default location most license scanners and developers look
first.

Furthermore, we've documented the copyright notices appropriately inside
the NOTICE file as specified by the Apache 2.0 license we're using.

* Add more stubs

* Add pyflakes

* Fix pyflakes

* Fix pyflakes for protocol folder

* Merge pyflakes and pycodestyle using flake8

* Remove PartyDB References

As a part of moving back over to Morten's repository we need to remove
the references to PartyDB.

The only reference remaining is in the README.md which relates to a
stubs repository we've setup. We need to decide what to do with that
repository when we open our request back to Morten's.

Related partydb/planning#1

* ..

* Small refactor

I've slightly refactored how the debug mode and monitor are turned on
and off inside the `bin/serve` script to make it clearer how these
variables get set.

* ..

* license and notice

* ..

* ..

* Add capcity to handle multiple inputs

* Simplify the client

* linting

* small changes

* correct website link