add a note to run tools/gcp commands from that directory
You get to this README via a link from the root README, but it's not
immediately obvious that you should switch to this directory to run
the commands.

For example, if you stay in the root and try to follow along, you will
fail on the second command

`gcloud compute ssh tfe-template < debian-install.sh` because `debian-install.sh`
will not be found.