Require Python 3.5 for development and Circle CI
In our setup.py we advertise that we supported Python 3.5, however, we
don't actually test whether we support 3.5 as a part of our continuous
integration suite.

Therefore, I've switched Circle CI over to 3.5 and updated our Makefile
to check that the callee is using 3.5 to test and publish our artifacts.

Closes #211