Private compare & LSB (#190)
Implements private compare and LSB from secureNN.

The major changes are the functionality in securenn.py.  As well as large amounts of supporting work to make the various tensors required to implement the protocol work with each other.  

Most of this supporting work revolves around knowing how to lift a tensor properly depending on the type.  For example, pond.py can now know that if it is dealing with a private pond tensor with a prime backing, calling prot.mul(tensor, 2) will lift the 2 into a prime backed pond constant.