Fix docker container so deploy task passes
The docker container was nolonger building because we forgot to update
the Dockerfile to be based off of Python 3.5.

I've updated the Dockerfile and added building the docker container to
our regular continuous integration task so we know it the container
builds before a pull request is merged into master.