Only show Keras API docs
Closes https://github.com/tf-encrypted/tf-encrypted/issues/587

Other docs are out of date right now.
This PR introduces the follwing:

* Started using [naploean](https://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html).
  In short, napolean generates restructured text from google-style docstrings.
* Only show the keras API

There is one thing that I am not happy with, and that is that the titles generated
by sphinx-apidoc are bad.  E.g. for the `tf_encrypted.keras` page, the page title will be
`tf_encrypted.keras package`.  Where something like `Keras` or `TFE Keras` would be preferable.

Apparently using `-M` in the apidoc command should use the module title instead, but this does
not appear to be happening.

https://github.com/sphinx-doc/sphinx/pull/6492/ has introduced a `--template` command, which
we should be able to eventually leverage to fix the above issue.  Whether we want to wait or
take a different approach with this is up to us.  I can also manually edit the .rst files
and change the header if we want.  I just wanted to get a discussion going.